[ADD/#1] SignUp 상태 설정
