+++ app/src/main/java/org/sopt/and/feature/signup/SignUpActivity.kt
@@ -1,0 +1,23 @@
+package org.sopt.and.feature.signup
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.setContent
+import androidx.activity.enableEdgeToEdge
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.tooling.preview.Preview
+import org.sopt.and.designsystem.theme.ANDANDROIDTheme
+
+class SignUpActivity : ComponentActivity() {
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        enableEdgeToEdge()
+        setContent {
+            ANDANDROIDTheme {
+                SignUpScreen()
+            }
+        }
+    }
+}
