+++ app/src/main/java/org/sopt/and/feature/signup/SignUpScreen.kt
@@ -1,0 +1,8 @@
+package org.sopt.and.feature.signup
+
+import androidx.compose.runtime.Composable
+
+@Composable
+fun SignUpScreen() {
+
+}
