+++ app/src/main/java/org/sopt/and/feature/signup/SignUpSideEffect.kt
@@ -1,0 +1,8 @@
+package org.sopt.and.feature.signup
+
+import androidx.annotation.StringRes
+
+sealed class SignUpSideEffect {
+    data class Toast(@StringRes val message: Int) : SignUpSideEffect()
+    data object NavigateUp : SignUpSideEffect()
+}
