+++ app/src/main/java/org/sopt/and/feature/signup/SignUpState.kt
@@ -1,0 +1,7 @@
+package org.sopt.and.feature.signup
+
+data class SignUpState(
+    val id: String = "",
+    val password: String = "",
+    val isButtonEnabled: Boolean = false
+)
